hw2 shell: kill to long working shells
It happens that a student could forget to process EOF supposed to
finish the shell process. In such a case usually the process
hangs. This commit makes shell working not more that 3 seconds.
In case if it works longer, it is killed and a warning is printed
with a piece of advice.