hw5 chat: clarify usage of data/udata
epoll_event.data.fd is almost never used in practice. Because in
practically all cases each client is not just a descriptor but
also a certain context, stored in a class/struct. Which makes it
hardly possible to use fd as user-data in the events. It would
require fullscan of the client array to match the fd to its client
when an event on that fd is raised.