hw1: introduce a new first homework
The old one didn't allow to use any of the interesting features
of coroutines, and their usage was rather unclear and random.

Lets replace the first homework with a totally new one where
coroutines are going to be the linchpin of the student solutions.

For that the coroutine engine had to be rewritten almost entirely
to introduce features like coroutine suspension and wakeup.

The homework now is about implementing a messaging system for the
coroutines to interact with each other.

This homework also had automatic tests and supports allcups
container testing.