hw4: clarify thread_pool_push_task() comment
It was complained about to be not clear enough regarding what to
do if the task is already pushed.

Since such behaviour is not possible to define properly, it is now
explicitly states to be "undefined behaviour". I.e. the
implementation can do anything with it (usually means do nothing
and don't consider such case at all).

Kudos to @abramchikd for noticing.